# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online image shop API",
        "description": "API za upravljanje porudžbinama, slikama, tehnikama slikanja, popustima i korisnicima kroz autentifikaciju i autorizaciju",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/popusti": {
            "get": {
                "tags": [
                    "Popusti"
                ],
                "summary": "Vraća sve popuste",
                "operationId": "65492d9ac96b3726e1bc4d29ebfdaaf2",
                "responses": {
                    "200": {
                        "description": "Lista svih popusta"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Popusti"
                ],
                "summary": "Kreiranje novog popusta",
                "description": "Kreira popust sa definisanim periodom važenja (dan/mesec od–do). Period ne sme biti duži od 31 dan. Podržani su popusti koji prelaze godišnju granicu (npr. 25.12 – 05.01).",
                "operationId": "619e6912a97efdc5594ddac71e95473e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "aktivan",
                                    "tip",
                                    "procenat",
                                    "danOd",
                                    "mesecOd",
                                    "danDo",
                                    "mesecDo"
                                ],
                                "properties": {
                                    "aktivan": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "tip": {
                                        "type": "string",
                                        "example": "praznik",
                                        "maxLength": 50
                                    },
                                    "procenat": {
                                        "type": "integer",
                                        "example": 15,
                                        "maximum": 100,
                                        "minimum": 1
                                    },
                                    "danOd": {
                                        "type": "integer",
                                        "example": 24,
                                        "maximum": 31,
                                        "minimum": 1
                                    },
                                    "mesecOd": {
                                        "type": "integer",
                                        "example": 12,
                                        "maximum": 12,
                                        "minimum": 1
                                    },
                                    "danDo": {
                                        "type": "integer",
                                        "example": 1,
                                        "maximum": 31,
                                        "minimum": 1
                                    },
                                    "mesecDo": {
                                        "type": "integer",
                                        "example": 1,
                                        "maximum": 12,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Popust kreiran"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    },
                    "422": {
                        "description": "Validaciona greška (npr. period duži od 31 dan ili neispravan datum)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/popusti-aktivni": {
            "get": {
                "tags": [
                    "Popusti"
                ],
                "summary": "Vraća sve aktivne popuste",
                "operationId": "11da2b8b48202a1364e2bd88d4336dbc",
                "responses": {
                    "200": {
                        "description": "Lista aktivnih popusta"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/popusti/{id}": {
            "get": {
                "tags": [
                    "Popusti"
                ],
                "summary": "Vraća jedan popust po ID-u",
                "operationId": "1fe5e46d5cf77c68cf57ef83f78bac44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Popust"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    },
                    "404": {
                        "description": "Popust nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Popusti"
                ],
                "summary": "Izmena popusta",
                "description": "Sva polja su opciona. Ako se menja period (danOd/mesecOd/danDo/mesecDo), moraju se proslediti sva četiri datumska polja. Period ne sme biti duži od 31 dan.",
                "operationId": "26cd038d50e0ffdd6d6832d2699e05e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "aktivan": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "tip": {
                                        "type": "string",
                                        "example": "praznik",
                                        "maxLength": 50
                                    },
                                    "procenat": {
                                        "type": "integer",
                                        "example": 20,
                                        "maximum": 100,
                                        "minimum": 1
                                    },
                                    "danOd": {
                                        "type": "integer",
                                        "example": 24,
                                        "maximum": 31,
                                        "minimum": 1
                                    },
                                    "mesecOd": {
                                        "type": "integer",
                                        "example": 12,
                                        "maximum": 12,
                                        "minimum": 1
                                    },
                                    "danDo": {
                                        "type": "integer",
                                        "example": 31,
                                        "maximum": 31,
                                        "minimum": 1
                                    },
                                    "mesecDo": {
                                        "type": "integer",
                                        "example": 12,
                                        "maximum": 12,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Popust izmenjen"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    },
                    "404": {
                        "description": "Popust nije pronađen"
                    },
                    "422": {
                        "description": "Validaciona greška (npr. nepotpuni datumski podaci ili period duži od 31 dan)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Popusti"
                ],
                "summary": "Brisanje popusta",
                "operationId": "0aa7d6e4eea3deec86638d6adcfc693d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Popust obrisan"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    },
                    "404": {
                        "description": "Popust nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}